Ponder and play with the idea for an RPN-inspired postfix stack-based programming language
Last night during dinner, I was preparing to implement an RPN calculator terminal utility in Haskell, when I started reflecting on how the limited subset of operators available on RPN calculators really could extend to any function. One thing led to another, and I spent the whole night thinking about such a language and how it could be implemented.

Such a language would appeal to my love of RPN, as well as give me a reason to implement a repl and compiler, both of which are highly enjoyable to me. I've always wanted to experiment with implementing closures, lambdas, and first-class functions in a language, and I can think of fun ways they would integrate with this postfix language.

I couldn't think of a great name, so I settled on Stakkelina, a portmanteau of 'Stack' and 'Javelina'. Such a name remains playful, unique, and (perhaps) memorable,  but also clues one into the general premise behind the language.

I'm going to very roughly digest and experiment with this idea by first implementing a REPL in Lua for it, and once I have the basics down, may write a compiler/transpiler that compiles that language into C as an intermediary language. I have lots of ideas for this, and I just want to dive in without thinking too much about.

I've uploaded some very meandering notepad scratchwork I typed up while pondering such a language last night, which contains visions of how I think it should look.

Next up, basic line parser in Lua, maybe a roadmap or rough outline of what I want.