Start writing lexer and structuring it out
I started writing a lexer collection of functions (I don't really feel like this needs to be a class, since its really just a series of function calls that don't need to be part of an instantiated object). Originally, I just started banging it out in Lua and was going to avoid Teal so as not to bog myself with learning that as I went along, but I found myself really gravitating towards wanting types, so I adopted Teal as the primary language for this.

Right now, what I've written isn't runnable. I'm about to start implementing the simplest functions with unit tests, so I want to commit before I start gutting anything. I also want to capture the lexer three-step outline in the top comment of the file.

Next up is to start implementing the tiniest functions, along with unit tests.