Find a way to get testing across scripts working in Teal, and write preliminary tests
Figuring out how to write tests in Teal was a bit interesting, because its restrictions on tables means if you declare functions in a file as local, its really hard to pass them to a separate test script in bulk, without just declaring them local in the original file. After a bit, I found a wait to do it by putting a helper function in the non-test file that is global, and when invoked by the test, makes global references to all of the needed functions available. This way, they won't pollute the environment during runs that aren't tests.

I had assumed that Lua and Teal in any given file, would read all of the top-level function names and note them before reading the bodies of any of the functions, but alas, I was wrong, and they do it in a single-pass down the file. I don't now how I used Lua this long without concretely knowing that (actually, I'm pretty sure I bumped into it before, but prob forgot). At any rate, forward declarations take care of the issue, allowing me to preserve the order that I want, at the cost of the extra lines. It's not so bad - it servers almost as a 'header file' of sorts.

My current unit test actually works and passes right now, though I still have to finish the other tests, as well as finish implementing more functions in lexer.

Next up, write several more tests, then move back onto writing more functions in Lexer.