Implement ordinary candidate builder and its corresponding test
I'm now starting to get to the core of the lexer, where candidate tokens are built. I wrote the first of three building functions, and then wrote a lengthy corresponding test for it, which currently passes.

Next up is to implement the remaining builder functions in Lexer, and write corresponding tests.