Implement candidate builder from quotation string and write test; safety commit
I implemented the candidate builder for tokens that start with a quotation mark (which was more tedious than I had assumed, probably because I'm tired), and built a lengthy corresponding test for it that tries many scenarios and edge cases.

I've finally got it to a point where it compiles, but is throwing an error on my third test in its function, where it is an empty quote. I'm pretty sure the solution is going tor require me chunking up and rearranging my code in the newly implemented function, probably going from a while to a do while loop, and since it took me so long up to this point, I want to do a safety commit first.

Next up is to refactor the quotation candidate builder until its test passes.