Complete quotation string builder and make sure test passes
I refactored the candidate builder for quotation mark delimited strings, and as suspected, a repeat until loop was a much better fit. Those unit tests I wrote really, really came in handy, and there was even a little bit of regression in the previous candidate builder that the tests caught beautifully.

The next logical thing to do is implement the candidate builder for apostrophe delimited strings. Its implementation should be nearly identical to the function I just completed, and so should its test. I will implement it first by copying the code, and should it wind up working with minimal changes, then it will be worth considering if they can be consolidated into one function via passing in some parameter, so as to reduce code duplication.

Next up, implement the apostrophe delimited candidate builder, and test it.