Complete candidate builder for apostrophe string and test it
Implemented the candidate token builder for the the apostrophe string function and its associated test by merely copying both the function and test for quotation mark delimited strings, and then swapping all references to quotation marks and apostrophes.

The test immediately passed. This proves what I mentioned in my previous commit - that the two functions can be consolidated if a delimiter is passed in as an argument. Committing before I attempt that.

It may be possible to also consolidate in the non-string delimited builder, but its logic is different enough that it may not feasible, or even if feasible, would be clunky or result in too many if-checks to be worthwhile.

Next up is to consolidate the two delimited string building functions if possible, and adjust the associated tests accordingly.