Consolidate the two candidate from string builders and adjust tests
I was successfully able to consolidate those two tests, as I had suspected. I thought updating the tests was going to be a chore of copying-and-pasting the two delimiters, but clever use of lambdas in both tests allowed me to avoid that entirely (lambdas rock). Both tests passed without issue.

I still think the consolidated method is ugly and longer than I would like, but I'm struggling to see how to break it into composite functions easily, since all the parts are connected.

Actually now that I think about it, I think I could do so with a little bit of elbow grease in the body of the repeat until loop, but frankly I just want to move on right now and am getting tired and bored of this part of the lexer, so I won't.

And for posterity, I did realize several commits ago that this probably could've been handled more gracefully and cleverly with regular expressions (which Lua doesn't have built-in, though they do have their own lesser 'equivalent' used in sub, find, and gsub), but I followed through because I wanted to have the experience of lexing at the character stream level at least once.

Next up is to implement the functions in lexer responsible for actually aggregating all of the lexeme candidates, and to write a corresponding test.