Complete segmenter and write test with supporting table comparison lambda
The segmenter was actually very close to being completed, save for a few errors in its code. A very long time was spent writing a comprehensive test, for which I just finished writing a supporting table comparison lambda to shallowly check for equality.

I just got done with a very brief test of that lambda, and it currently passes. Now is to actually utilize that lambda to assert over all the string and result pairs I made in the test.

Next up is to finish the test for the segmenter and debug as the test reveals errors, which it undoubtedly will.