Test lexical segmenter until it passes
After much iterating and bug-chasing, I finally have it so that all twenty cases pass in the segmenter test I wrote. That test was amazing, and allowed me to squash so many bugs so quickly. It even was right when I thought it wasn't, catching things that I didn't realize I put in there (like stray invalid apostrophes).

The segmenter is thus fully operational now, so part one of three of the Lexer is complete. The two remaining parts are the lexical validation step and the tokenization step, which should be easier than what I just did, albeit they will start requiring me to make design choices about the key operators and functions that I want included in this language.

Next up is to start writing functions to support the lexical validation phase. Some scratchwork and prototyping might be in order. It's also worth considering if these phases should be broken into files (which is unclear, because the Lexer is one 'thing', but it's a collection of functions and not a class because these are one-pass actions that don't require persistence).