Returned on a whim - throw uncommitted and existing work into scrap heap
Out of nowhere (well, actually having been enticed by wanting to implement a PowerShell or Assembly Lisp/Scheme), I'm returning to this (if for only a moment) to scratch my compilers and interpreters itch.

Let's see if I can get something going, lickety-split. We'll do this in Fennel, and if I'm too lethargic, then Lua. Yay, reinventing the wheel.