Start implementing basic functions needed for lexer, tokenizer, parser, and interpreter
Wrote an assortment of micro-functions to be used for lexing - while I have long since learned how to use LPEG, I feel it would be overkill here, since this language should be fairly easy to parse character-by-character, sparing the project from having to have a dependency. I'd like to stay in-source as much as possible.

Next up is to start implementing functions that actual lex, segment, and validate/parse a char stream.