Implement function for finding next token
Wrote a simple function that finds the next token in a stream, given that the start of the stream is beyond its current position.

Instead of using sub and match to continuously query strings or reparse them again and again, I have opted to make the character 'stream' simply be a string with an index that updates as one moves across it. This stops duplication issues where each little change causes a new string (which can happen in Lua) and simultaneously encourages good programming habits.

Next up is to continue to steam towards getting a simple parser / interpreter going.