Implement function for reading a token at a given position
I now have both a function to find the position of the next token and a function to extract / read a token at any given position - this is the smallest viable lexer! It has been tested via REPL to be functioning.

Next up is to implement several functions - one that lexes / segments the entire character stream into a list, another that lazily produces tokens when invoked, and of course, a function for evaluating tokens and functions for validating lexemes / segments.