Implement lexer as a token-iterating closure
I now have a fairly elegant and simple lexer/tokenizer that is implemented in the form of a self-iterating closure. It has been tested via REPL at length - including a fair number of edge-cases - and it works exactly as desired.

It is - of course - a preliminary version that does not seek to cover all bases and cases - I can make it more thorough when and if it gets to that point.

Next up is to begin validating the lexemes into tokens, and then parsing them into an in-memory abstract syntax tree (AST).