Refactor project file structure; rename files and functions to match actual language processing correspondence
As I started to dive deeper into how I was going to implement the nitty-gritty of the parsing, I was forced to think more specifically about the relationship between lexing, tokenizing, parsing, interpreting, compiling, and language processing in general. I've thus renamed and restructured my project accordingly to more accurately match what each unit will be doing and what process they correspond with.

I've tested it via REPL to ensure it still works as desired. There will of course be more function-shuffling between files, as some of these functions below in a tokenizer rather than a lexer. Also, it might be best to move some functions to a common include, rather than misleadingly keep them in a unit that they are actually abstracted from (but still used in).

Next up is to start writing tokenization functions and parsing functions -- this can be done concurrently, as single-pass and multi-pass processing are both viable and legitimate options. Just have to keep the momentum up.