Implement token iterator/streamer
In a manner similar to the lexer-iterator, I implemented a token-iterator. It was actually much easier, as there were no variables to manage this time due them being encapsulated in the lexeme stream already.

I've tested it, and the token streamer works well. However, it does not handle nil/end-of-stream well and errors out.

Next up is to get the token-iterator to handle the stream being exhausted properly.