Implement testing framework and unit tests
Due to mental fatigue from lack of sleep, I struggled with what should have been fairly straight-forward. I set out to fix the token-iterator so that it could handle end-of-streams correctly, but wound up with other errors.

Tired of manually typing everything out again and again in REPL when wanting to correct one thing, I decided to start writing unit tests and to outline a simple unit test framework for the project. The tests currently work, but they print to stdout instead just performing their work and reporting the results.

It would be worthwhile to tweak the framework so either one could be chosen - test until halt, and test each module, giving a report of which ones failed and which ones didn't.

Next up is to continue the tokenizer and parser, creating tests as needed.