Implement basic REPL
I realized that I may not need an AST, at least not now, and could suffice with a stack and some metadata alone for now (compiling would be difference). Really, I realized that I had to jump up to the REPL and start working back 'down' to the interpreter and parser.

The REPL loop works well for now, though it doesn't make use of any features yet.

Next up is to work on the REPL, interpreter, and parser, as each specific task sees fit.