Start writing interpreter and built-ins
As I thought about parsing and evaluating, I began to realize in a stack-based language that they can be the exact same thing. However, as I plan on extending this with a transpiler or compiler, there will have to be a separate parsing phase at some point, even though a traditional AST isn't need in this type of program.

Nonetheless, I must just start implementing and see what works.

Nothing new here is runnable, but I'm committing to keep track of changes incrementally. There will be lots of implementation coming up that can't really be tested until it is synthesized.

Next up is to keep moving towards a running interpreter.