Extend lexer test to lex-one; sort out pretty-printing
Successfully extended the tests to include lex-one, and was able to sort out how to use Fennel's built-in pretty-printing, so that I don't have to reinvent something similar.

Next up is to continue to write tests (and should move towards more traditional ones), and continue refactoring the lexer, moving on afterwards to tokenizer and parser and interpreter.