Finish implementing lex-all, completing the lexer
I refactored lex-one just a bit (since that mutating set on the table need not be in the let declaration list), and then completed lex-all using the functions written last commit. After a bit of REPL debugging, I now how a solidly working lexer!

It is as complete as it can be, so I'm free now to move onto the higher levels to do something very similar (giving options for ingesting one unit, ingesting all units, or providing a lazy producer).

Next up is to implement similar ingestion options for the tokenizer and parser, to be used by the interpreter as needed.