Refactor parser to be increase readability
My previous implementation, while identical in functionality as what I just wrote, was an absolute abomination. What I have now is so much cleaner and nicer to look at. The outcome of the primitive test remains unaltered (successful).

Next up is to implement that various singular and bulk processing methods in parser as was done in lexer and tokenizer. Then tests should be written.