Implement remaining parsing function by repurposing tokenization functions
I implemented a variety of parsing functions by simply copying over the tokenization functions and changing the names of certain symbols. This tells me that there is are underlying abstractions that can be pulled out (which are obviously iterators, delegators, and map, in this case) to minimize code duplication. However, I'm not going to bother right now, as the reuse is so minimal that implementing the abstractions and building off top of them would result in almost the same number of lines of code, or perhaps marginally fewer. I'm not expecting to need to reuse them beyond what was already implemented (though of course I may in other programs - I'll cross that river when I come to it.

The tests I wrote still all pass, including the one for the parser, but it has not been extended yet to include testing of these newly implemented functions.

Next up is to write tests for all of the remaining parsing functions.