Implement tests for remaining parsing functions
In much the same manner as last commit, I copied the tests from tokenizer and repurposed them for use with parser, adapting them as needed. All tests passed after I did some debugging, but it's worth noting that I've only implemented a subset of the token types and node types that need to be implemented. It passes for that limited subset, though.

Next up is to start working on the interpreter, and to try to get simple evaluations working as soon as possible. It is then desirable to go back and implement the remaining token and node types.