Implement stack data structure and associated functions
Since the interpreter will clearly need a stack as it evaluates REPL commands and programs, I went ahead and implemented one. Nothing remarkable about it.

It was tested via REPL and works as desired.

I thought of adding the ability to handle varargs in create-stack and push, but I don't know that I'll need such functionality, and if I did, I would have to move the vector argument to the first position, which would annoying since I like to have the elements passed first(representing e into v).

Next up is to return to the interpreter and continue trying to write it or whatever it needs.