Start implementing evaluation functions in the interpreter
I started to write the evaluation functions in the interpreter, but stopped to ruminate on precisely how I should move forward. I have determined I need a symbol table module, as well as a frame module. There are more things I need but those are the best to start with.

Next up is to implement either the symbol table module or the frame module.