Attempt to adjust path where Fennel searches for modules
I started to skeleton out the needed files for going forward, and then took a long break to work on macros in another repo that I could eventually bring over here (I succeeded, but need to polish them up first). I also am bothered by how I had to include the tests folder inside the src folder due to path recognition issues, so I wanted to figure out how to add a directory and its subdirectories recursively via a command in a script.

I just succeeded in 'theory.fnl' doing that, showing proof of concept.

Next up is to restructure my project so that the tests are in the top-level tests directory, and adjust my runner accordingly. Then to continue with frame and symtable, as well as with bringing over the import macro I'm working on elsewhere.