Refactor directory structure to move tests to top-level tests directory
I struggled with this more than is reasonable, do to the (for me) janky Lua path searching mechanics. I eventually sorted it out and got the test runner to locate all needed files from the top-level execution of the script.

Next up is to continue with writing frame and symtable.